% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_chd_age.R
\docType{data}
\name{export_chd_age}
\alias{export_chd_age}
\title{CHD - age data}
\format{
A data frame with 7772 rows and 2 variables:
}
\source{
{HjerteTal data}
}
\usage{
export_chd_age
}
\description{
A dataset
}
\keyword{datasets}
