% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_chd_sex.R
\docType{data}
\name{export_chd_sex}
\alias{export_chd_sex}
\title{CHD - sex data}
\format{
A data frame with 7772 rows and 2 variables:
}
\source{
{HjerteTal data}
}
\usage{
export_chd_sex
}
\description{
A dataset
}
\keyword{datasets}
