% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_diag.R
\docType{data}
\name{export_diag}
\alias{export_diag}
\title{Diag data}
\format{A data frame with 7772 rows and 2 variables:}
\source{
{Aalborg census 1850. Data entered by ___}
}
\usage{
export_diag
}
\description{
A dataset containing the age and gender of every individual recorded in 1850
census.
}
\keyword{datasets}
